pyproject.toml
[project]
name = "langchain-doc-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "black>=25.1.0",
    "certifi>=2025.8.3",
    "fastapi>=0.116.1",
    "firecrawl-py>=4.3.6",
    "isort>=6.0.1",
    "jinja2>=3.1.6",
    "langchain-chroma>=0.2.6",
    "langchain-community>=0.3.29",
    "langchain-pinecone>=0.2.12",
    "langchain-tavily>=0.2.11",
    "langchainhub>=0.1.21",
    "langsmith>=0.4.27",
    "nltk>=3.9.1",
    "openai>=1.107.2",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "streamlit>=1.49.1",
    "streamlit-chat>=0.1.1",
    "tavily-python>=0.7.12",
    "tiktoken>=0.11.0",
    "tqdm>=4.67.1",
    "unstructured>=0.18.14",
    "uvicorn>=0.35.0",
]
